Save zenup files to particular name
When the zenup files were being downloaded they were being saved as 'download'
because the url ended in download instead of being saved as the proper filename.
Added the -O option to wget so the files were saved properly.

When the files were saved as 'download' they were not being installed because
yum and cp were running with the actual filename but the files were saved as
'download'.